+++ internal/util/diff_util/operation.go
@@ -1,4 +1,4 @@
-package utils
+package diff_util
 
 import "strings"
